New Module, New Client, New API
Wrote a simple module to manage a minecraft server run in the server
directory, and added multiple commands to assist with operation, though
this module could be expanded to work with any executable file in the
execution environment. Also created/converted the server into a dll for
use when making modules so as to provide access to the UDPFrame class,
though it's since been expanded to provide many functions to the modules
directly. As such, the reflection code is moving from reflectively
finding all hook functions, to finding a single function that allows the
module to hook directly to the events in the Program class. Other stuff
was added to the server to clean up and assist with the migration to dll
delivery of items to the modules. Also reworked the client to allow
console commands to be sent from the client to the server and the
response sent back to the client. More features will be included later,
including some authentication system.