Slight Additions to MCServer
Added checks for status of server as a command, and to avoid errors in
starting and stopping the server. There still isn't a hard check for the
status of the process when executing to inputstream to start or stop,
but the pause statement on the launch bat should prevent issues with
server stopping internally and on the wrapper together; the only way to
officially stop the server is via the terminal commands. Next up is
creating a system to relay the server output to the client upon request,
though that needs a packet hook, not just commands.