Better Client Handling
Changed all client dictionaries and lists to use the client UID instead
of IPEndPoint, as the endpoint could potentially change. Added new
functions to update the IPEndPoint directory for each packet if needed
to keep the client directory up to date. Also added hearbeat packets to
ensure that UDP Hole Punching is always available for the client to
recieve passthrough packets. Also added the ability for clients to send
packets to other clients based on uid, though this is impractical. Still
haven't added packet verification to prevent drops, nor end-to-end AES
encryption.