Chat Rooms
Added a really terrible chat room module (biggest problem is the lack of
client identification with nicknames or something). Also fixed the
client to ignore output from a command if the data in UDPFrame is null,
so commands don't have to return a blank line when no output is desired.