Updated Module loading and Interface Dependency
Added an interface that all modules can implement for their main class
to assist with finding the methods to add the hooks. Also toyed around
with appdomain loading of modules, but it will not work for this system,
so everything is loaded into the default domain. Switched the loading of
modules to byte-wise so that the binaries can be deleted during runtime
and replaced. Also added a helper class for the modules to abstract the
methods used by them a bit. Started with the authentication module for
the user system, but have decided to keep it in the main server program
to avoid excess communication channels being made.